+++ smoke-tests/framework/src/main/java/com/microsoft/applicationinsights/smoketest/fakeingestion/ProfilerState.java
@@ -7,5 +7,7 @@ public enum ProfilerState {
   unconfigured,
   configuredEnabled,
   configuredDisabled,
-  manualprofile
+  manualprofile,
+  targetedMatching,
+  targetedUnmatched
 }